revert uneeded changes
