+++ doozer/doozerlib/util.py
@@ -222,6 +222,23 @@ def get_docker_config_json(config_dir):
         raise FileNotFoundError("Can not find the registry config file in {}".format(config_dir))
 
 
+def isolate_git_commit_in_release(release: str) -> Optional[str]:
+    """
+    Given a release field, determines whether it contains
+    .git.<commit> information or .g<commit> (new style). If it does, it returns the value
+    of <commit>. If it is not found, None is returned.
+    """
+    match = re.match(r'.*\.git\.([a-f0-9]+)(?:\.+|$)', release)
+    if match:
+        return match.group(1)
+
+    match = re.match(r'.*\.g([a-f0-9]+)(?:\.+|$)', release)
+    if match:
+        return match.group(1)
+
+    return None
+
+
 def isolate_pflag_in_release(release: str) -> Optional[str]:
     """
     Given a release field, determines whether is contains
